refactor: crud

+++ currency-and-account-service/src/main/java/com/example/bank/dto/CreditResponseMessage.java
@@ -5,13 +5,14 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
-public class CreditResponseMessage implements Serializable{
+public class CreditResponseMessage implements Serializable {
 
     private String accountNumber;
 
-    private Long money;
+    private BigDecimal money;
 }
