+++ transaction-service/src/main/java/com/example/bank/request/DrawMoneyRequest.java
@@ -5,13 +5,16 @@ import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 @AllArgsConstructor
 public class DrawMoneyRequest extends ApiRequest{
+
     @NotEmpty(message = "INVALID_ACCOUNT_ID")
     private String sourceAccountNumber;
 
     @NotEmpty(message = "INVALID_MONEY_INPUT")
-    private Long money;
+    private BigDecimal money;
 }
